Preferences fix (hacky)
I can't seem to find any other way to get the new settings (Stereo, Fade, and FPS) to save properly except when I do this. If someone can edit it so it saves it correctly, that would be great.
The bug right now is that once you change Fade or Stereo settings, you can't change it back without deleting the settings.onbs file.