NBS Format Additions (Loop, MIDI, Lock)
Fixed MIDI Filename not being saved to NBS file.
Added a preference option to loop till end of bar (if false, loops at the last tick with a note)
Added a maximum amount of times a song can loop (0 = infinite)
The state of a layer being locked is now saved to the NBS file.
Some GUI graphic fixes.