Store settings as .ini file rather than binary
This prevents conflicts between versions. The order in which data is written to the file doesn't matter, entries can be added or removed anywhere without causing issues. If a key doesn't exist, it remains as default; if a stray key is found it's simply ignored. This means that a lot of unused variables (that needed to exist just to fill blank spaces in the file) were also removed:

- Removed soundsystem, settings_verloaded and settings_version variables
- Removed variables related to the MIDI export feature which were removed long ago
- rhval is no longer stored in the settings