Fix wrong menu actions in key edit mode
Since there are two fewer menu items on the key edit mode, the actions following 'Decrease key' were all incorrect. The code now makes up for that by adding 2 to 'sel' (the number of the menu item clicked) when it is greater than that value and when editing the key.