Fix scrollbar being drawn outside the screen area
The calculation for the amount of columns wasn't taking into consideration the width of the scrollbar, which, in specific window resolutions, would cause it to be rendered too far to the right - thus making it inaccessible.

To prevent this, I'm now subtracting half of the width of the scrollbar (8). Subtracting the full width (16) would cause too much useful space to be wasted just so the scrollbar is fully rendered (a full column could be drawn in the space it occupies). Subtracting *half* of the width, half of the width is as hidden as it can get.