Fix emittters lasting much longer than they should
For some weird reason the audio_get_length() GM function seems to return a value that's consistently 4-5x the actual length of the sound. On top of that, since this value is multiplied by the inverse of the sound pitch to make up for the change in duration, it was the case that some lower pitched sounds could stay 20 seconds in memory instead of 4, causing the polyphony limit to be reached much quicker.

The fix was to simply divide the length by 4, which is just enough to prevent cutoffs in the lower notes. (I found the lowest GM length/actual length ratio in the default instruments to be 4.09, so it may be a bit longer than the sound but will never be shorter.)