Prevent marker from skipping notes on fast playback speeds
The cause of this issue was that the marker only plays notes that it flies over. When the playback speed is fast enough (or when the program is lagging), sometimes it will move more than a whole tick, thus having the chance of not stopping for at least a single frame in some ticks.

The fix for this issue was, rather than playing only the ticks that the marker "touches", to play every tick after the previous one that was played, up to the current one. This makes sure no ticks are missed, even though they may clip sometimes.

This absolutely fixes #101. Unbelievable!