Update converted project with the most recent changes
Forgot to pull newer commits before converting the project. Oops.

(This is just the merged resource pack from PR #172 and additional data files which GameMaker no longer uses. They were present in the master branch from which the PR was created, but not on the development branch. The merge (be7fbea40c3b158bcbb69fc9e3c42fe0e95fbfed) then added those files to the development branch, but those changes had not been pulled as the project was converted, so the support files were not deleted on the conversion as they weren't there to begin with).