Fix hardcoded rule preventing editing most instrument name boxes
There was a hardcoded rule in the text_edit code from when layer IDs were still >100. I have no clue why this is the case, but something may be implemented differently in the layer name boxes than the other text boxes. In any case, this is something else to keep in mind in case IDs have to be rearranged again. >_>