Prevent wallpaper from being copied if transparency is not enabled
Only copies the wallpaper in either one of those instances:

1) the user starts the program with the Fluent theme AND transparency enabled;
2) the user enables the Fluent theme having transparency enabled;
3) the user enables transparency having the Fluent theme enabled.

This prevents the wallpaper from being copied unnecessarily for users with transparency disabled, avoiding potential issues and allowing users to disable it in case of any problems (i.e. they can manually set acrylic=0 in settings.ini). It also has the added bonus of updating the wallpaper without needing to restart the program.