Fix song info window not opening in player mode when song has no title
Fixes #276

NBS makes a distinction between when the 'Song info' window opens automatically at the start of the song (`window = w_songinfo`) vs. when you open it manually by going to Settings > Song stats (`window = w_songinfoedit`), in order to show the "Click anywhere to dismiss" text or the 'Edit' and 'OK' buttons accordingly.

However, to prevent the song stats from appearing when the song is loaded if it has no title, a check was being done inside the window draw function itself, which returned if the song name was empty AND the window was opened in "no edit mode" (`window = w_songinfo`).

The thing is: since in player mode you're not supposed to edit the song as well, the code for it took advantage of the existing "no edit mode" and used it for opening the window from the menu as well (whereas it was originally only intended to be used when opening the song, so the starting popup would be skipped when appropriate). The side-effect of this is issue #276: the window doesn't appear at all in player mode if the song has no title.

To fix this, I moved the check for an empty song title to the `load_song` function, which makes more sense than checking this every time the window opens -- since it's the only case where we want to make sure the song title is not empty.