Fix crash on schematic/datapack exp. due to old custom instrument limit
Even though the custom instrument count limit was increased to 240, the schematic export settings for what block to place under custom instruments was still taking the old limit into account (18 custom, 34 total), causing all sorts of crashes when reading from this array (`sch_exp_ins_block` and `sch_exp_ins_data`).

While one would expect this to only affect the schematic export, there's one completely unrelated usage of this array in the code: checking which block to spawn on the data pack export visualizer.

This issue was causing crashes when exporting data packs with the visualizer enabled, which is what led me to discover it; however, it likely also happened on schematic export.