Fix notes barely outside of range being considered in range
The previous way of detecting out of range notes would consider anything greater than 32 and less than 58 in range -- which meant values like 32.5 or 57.5 were considered in range, when in practice they're already out of range.

For instance, a G5 tuned -80 cents would turn out to be:

58 - (80/100) = 58 - 0.8 = 57.2

which was considered in range before, but is *actually* out of range, since it's 20 cents sharper than the highest note Minecraft can support (57).

This caused notes that lied within this boundary (`32 < key < 33`, `57 < key < 58`) to have the incorrect prefix appended to it, causing discrepancies during playback.