Fix version comparison not taking two-digit minor versions into account
This would break when 3.10.0 is released, as it would compare only the third character of both, resulting in 1 < 9.

Any different version string is now considered to be a different version. (With the possibility to download pre-release versions, the user might be downgrading from a pre-release version to a stable version)