Fix popup drawing code always relying on alpha value of `obj_popup`
The reason for moving the popup drawing code to a separate script was to be able to call it externally, without needing to create a `popup_obj`. Reading the alpha value from the `popup_obj` without it existing would likely cause a crash (hence the `try`/`catch` block added in 747f627d853fb9fb6d0a78b4834e0e7deb7072b5). As such, the most appropriate setup is to pass the popup's alpha value as an (optional) argument, assuming a default value of 1 when the function is called externally without providing that argument.