Fix update check failing at random
Although this code contains many optimizations, the "actual" fix is on line 26 of `check_updates.gml`: upon doing an HTTP request, GameMaker sends multiple callbacks to inform async events about the "progress" of the request, during which the `status` field on `async_load` will be 1. The `result` will only be available when `status == 0` (success), so we must ONLY stop processing the callbacks (setting `update_http` to -1) when this occurs.

Previously, the script was setting `update_http` to -1 as soon as the first callback was received, which means `result` might not have been available yet. The `http_status` would also be `undefined` when this occurred, which meant that checking for the update would randomly fail.

Read more about it here: https://manual.yoyogames.com/GameMaker_Language/GML_Reference/Asynchronous_Functions/HTTP/http_get.htm
https://docs2.yoyogames.com/source/_build/2_interface/1_editors/events/async_events.html