Fix "Sucessfully updated" text not appearing after updating
Create a new variable `update_success` that bypasses the other status options when true. The "Successfully updated!" string would never actually appear, since the async update check would override it later with "You are using the latest version!" (`update == 2`).

This text only started appearing before due to a bug solved in 94405482529b1b585de1a570092d92fbfa858cb0, in which the update check would be stuck at "Checking for updates...". Since the logic for actually verifying if there's a newer version was never executed, the `update` status "had a chance" to stay set to 3, which would display the text properly.